% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/served-data.R
\docType{data}
\name{served}
\alias{served}
\title{Person Served}
\format{An object of class \code{data.frame} with 426 rows and 26 columns.}
\usage{
data(served)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(served)
}
\keyword{datasets}
